Refactor compilers using traits
